Fill in data import as supplied by @NoYo25
This was supplied over Discord because the other file was too large to
add on Windows without git-lfs or we're not entirely certain how to do
this on such a box, likely possible.  It's the second iteration of that
file and this was the fastest route to getting it in the stack.