Import legislation
This is a somewhat manual conversion (using macros in a text editor).
There's a good shot most of this is correct but some manual processing
was involved too.

References to the legislation would be extremely welcome but this is the
closest I'll get right now.  Check
https://codex.opendata.api.vlaanderen.be:8888/sparql as a good starting
point to add these references 🤓