Add isDefinedBy links for legislation
I know, we should refer to the legislative entity.  This does define the
resource but we could follow the model better by searching through the
sparql endpoint of the Vlaamse Codex.  I would be surprised if this
would be the resulting URI but we did not search so it might be.